added exception to the code to sole problem o giving music file that alredy exists

+++ app.py
@@ -16,12 +16,17 @@ class Youtube:
         out_file = video.download(output_path=path)
         base, ext = os.path.splitext(out_file)
         new_file = base + '.mp3'
+        
+        # Check if the file already exists
+        if os.path.exists(new_file):
+            error_message = 'Error: File already exists.'
+            return error_message
+        
         os.rename(out_file, new_file)
         return new_file
-    
-    
 
-y = Youtube()
+    
+y=Youtube()  
 
 @app.route('/', methods=['GET', 'POST'])
 def index():
@@ -29,7 +34,20 @@ def index():
         link = request.form['link']
         path = request.form['path']
         file_path = y.video_downloader(link, path)
+        
+        # Check if file_path is an error message
+        if isinstance(file_path, str):
+            return '''
+                <form method="post">
+                    Enter youtube video link: <input type="text" name="link"><br>
+                    Enter the path to download the file: <input type="text" name="path"><br>
+                    <p>{}</p>
+                    <input type="submit" value="Download">
+                </form>
+            '''.format(file_path)
+        
         return send_file(file_path, as_attachment=True)
+    
     return '''
         <form method="post">
             Enter youtube video link: <input type="text" name="link"><br>
@@ -37,6 +55,7 @@ def index():
             <input type="submit" value="Download">
         </form>
     '''
+
 
 if __name__ == '__main__':
     app.run()
